requirements.txt
Flask==3.*
functions-framework==3.*
serverless-wsgi
google-cloud-pubsub==2.*
python-dotenv